fix: stop re-exporting types through "use server" boundaries
Four Wave 3 server-action files re-exported a type via
`export type { X }`. Next's server-action transform catches the
re-exported name in the module's export list and emits a runtime
reference to it — but the name is a type with no value, so the SSR
chunk throws `ReferenceError: X is not defined` on module evaluation
and crashes the affected routes.

Move every such type to its plain `src/lib/*` module and have consumers
`import type` it from there directly. `reorder`, `utilization`,
`reservation-conflicts` and `project-costs` server files no longer
re-export or hand types to consumers. `UtilizationSummaryRow` moves
from `src/server/utilization.ts` to `src/lib/utilization.ts`.

Local `export interface`/`export type` declarations are unaffected —
only re-export specifiers trip the transform. Documented the rule in
CLAUDE.md so it can't recur.
